+++ sdk/keyvault/azure-security-keyvault-keys/CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### Bugs Fixed
 
+- Fixed challenge-based authentication to correctly parse the tenant ID from DSTSv2 authority URIs.
+
 ### Other Changes
 
 ## 4.11.2 (2026-08-18)